Crate gitup

This program pulls the main branch of the repo from which it’s called, then deletes any local branches that are not ahead of main, and finally checks back out the original branch. The main branch defaults to main, but master is used as a fallback if no main branch is found.

TODO

Delete remote branches behind trunk.

Don’t mess up “checkout -” by checking out main. I tried not checking out main, but after fetch –prune, the user still sees a list of obsolete branches the next time they pull –prune; so now this program checks out main just so it can run pull –prune per se.

Support squashed merges.

Structs

Args 🔒

Policy for which branches to delete.

Enums

Error 🔒

Constants

TRUNKS 🔒

Possible names of local trunk branches, in order of preference.

Functions

format_git_command 🔒

git 🔒

git_loud 🔒

is_working_copy_clean 🔒

local_trunk 🔒

Returns the local main branch (trunk) name, or an error.

main 🔒

main_imp 🔒

run_git 🔒

Returns (success, stdout, stderr).

trim_branches 🔒

Splits the specified Git output into lines, excludes any line beginning with “*”, and trims leading whitespace from each line. Note that the output is not necessarily a list of simple branch names; e.g., if the output is from git branch --verbose.

upstream 🔒

Type Aliases

Result 🔒