Crate gitup

This program pulls the main branch of the repo from which it’s called, then deletes any local branches that are not ahead of main, and finally checks back out the original branch. The main branch defaults to main, but master is used as a fallback if no main branch is found.

Structs

SimpleError 🔒

Functions

format_git_command 🔒

git 🔒

git_loud 🔒

is_working_copy_clean 🔒

local_trunk 🔒

Returns the local main branch (trunk) name, or an error.

main 🔒

main_imp 🔒

run_git 🔒

Returns (success, stdout, stderr).

upstream 🔒